backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.18.0
scikit-learn>=1.4.0
scipy>=1.11.0
google-generativeai>=0.3.2
python-multipart>=0.0.7
fpdf>=1.7.2
openpyxl>=3.1.2
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-experimental>=0.0.49
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
# --- NEW: RAG & SQL Dependencies ---
chromadb>=0.4.22
pypdf>=4.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0           
psycopg2-binary>=2.9.9